move associate modal and disassociate button up to components for use across screens
