+++ awx/ui_next/src/components/AssociateModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './AssociateModal';
