+++ awx/ui_next/src/components/DisassociateButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './DisassociateButton';
